test(NODE-4999): Got repro working
